+++ spec/examples/double_with_type_as_string.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+class DoubleWithTypeAsString < Actor
+  input :value, type: [Integer, 'Float']
+  output :double
+
+  def call
+    self.double = value * value
+  end
+end