+++ spec/examples/increment_value_with_rollback.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 class IncrementValueWithRollback < Actor
-  input :value, type: 'Integer'
-  output :value, type: 'Integer'
+  input :value, type: Integer
+  output :value, type: Integer
 
   def call
     self.value += 1