+++ spec/examples/play_actors.rb
@@ -7,9 +7,9 @@ require_relative './set_name_to_downcase'
 require_relative './increment_value'
 
 class PlayActors < Actor
-  input :value, type: 'Integer'
-  output :value, type: 'Integer'
-  output :name, type: 'String'
+  input :value, type: Integer
+  output :value, type: Integer
+  output :name, type: String
 
   play IncrementValue,
        DoNothing,