+++ spec/examples/set_name_to_downcase.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 class SetNameToDowncase < Actor
-  input :name, type: 'String'
-  output :name, type: 'String'
+  input :name, type: String
+  output :name, type: String
 
   def call
     self.name = name.downcase