+++ spec/examples/set_wrong_type_of_output.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class SetWrongTypeOfOutput < Actor
-  output :name, type: 'String'
+  output :name, type: String
 
   def call
     self.name = 42