+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Added:
 - All errors inherit from `ServiceActor::Error`.
 - In case of argument errors, raise an `ServiceActor::ArgumentError` instead of
   a `ArgumentError`.
+- Allow classes as well as strings in type definitions.
 
 Fixes:
 - Allow inputs and outputs called `before`, `after` and `run`.